monkey-pic/scripts/ai3d/requirements.txt
torch
opencv-python
numpy
open3d
pymeshlab
pillow
scipy
